+++ genai-perf/tests/test_input_rankings.py
@@ -28,12 +28,12 @@ from pathlib import Path
 from unittest.mock import mock_open, patch
 
 import pytest
+from genai_perf.inputs.file_input_retriever import FileInputRetriever
 from genai_perf.inputs.input_constants import (
     ModelSelectionStrategy,
     OutputFormat,
     PromptSource,
 )
-from genai_perf.inputs.input_retriever_factory import InputRetrieverFactory
 from genai_perf.inputs.inputs import Inputs
 from genai_perf.inputs.inputs_config import InputsConfig
 
@@ -77,8 +77,8 @@ class TestInputsRankings:
             batch_size=batch_size,
             num_prompts=100,
         )
-        input_retriever_factory = InputRetrieverFactory(config)
-        dataset = input_retriever_factory._get_input_dataset_from_rankings_files(
+        file_retriever = FileInputRetriever(config)
+        dataset = file_retriever._get_input_dataset_from_rankings_files(
             queries_filename=queries_filename, passages_filename=passages_filename
         )
 
@@ -99,7 +99,7 @@ class TestInputsRankings:
             match="Batch size cannot be larger than the number of available passages",
         ):
             config.batch_size = 5
-            input_retriever_factory._get_input_dataset_from_rankings_files(
+            file_retriever._get_input_dataset_from_rankings_files(
                 queries_filename, passages_filename
             )
